+++ lib/view_models/stream_view_model.dart
@@ -86,7 +86,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
   /// Fetches the thumbnail for a live stream.
   /// Parameters:
   ///   [streamId] - The identifier of the stream.
-  Future<String> fetchStreamThumbnail(Int64 streamId) async {
+  Future<String> fetchStreamThumbnail(int streamId) async {
     try {
       _logger.i('Fetching thumbnail for live stream ID: $streamId');
       return await StreamHandler(_grpcHandler).fetchThumbnailStreams(streamId);
@@ -99,7 +99,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
   /// Fetches the thumbnail for a recorded stream.
   /// Parameters:
   ///   [streamId] - The identifier of the stream.
-  Future<String> fetchVODThumbnail(Int64 streamId) async {
+  Future<String> fetchVODThumbnail(int streamId) async {
     try {
       _logger.i('Fetching thumbnail for VOD stream ID: $streamId');
       return await StreamHandler(_grpcHandler).fetchThumbnailVOD(streamId);
@@ -109,7 +109,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
     }
   }
 
-  Future<void> fetchStream(Int64 streamId) async {
+  Future<void> fetchStream(int streamId) async {
     _logger.i('Fetching stream');
     state = state.copyWith(isLoading: true);
     try {
@@ -133,7 +133,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
     }
   }
 
-  Future<void> fetchProgress(Int64 streamId) async {
+  Future<void> fetchProgress(int streamId) async {
     state = state.copyWith(isLoading: true);
     try {
       final progress =
@@ -148,7 +148,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
     }
   }
 
-  Future<Progress> fetchProgressForStream(Int64 streamId) async {
+  Future<Progress> fetchProgressForStream(int streamId) async {
     try {
       final progress =
       await StreamHandler(_grpcHandler).fetchProgress(streamId);
@@ -158,7 +158,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
     }
   }
 
-  Future<void> updateProgress(Int64 streamId, Progress progress) async {
+  Future<void> updateProgress(int streamId, Progress progress) async {
     _logger.i('Updating progress');
     state = state.copyWith(isLoading: true);
     try {
@@ -170,7 +170,7 @@ class StreamViewModel extends StateNotifier<StreamState> {
     }
   }
 
-  Future<void> markAsWatched(Int64 streamId) async {
+  Future<void> markAsWatched(int streamId) async {
     _logger.i('Marking stream as watched');
     state = state.copyWith(isLoading: true);
     try {
