+++ lib/views/chat_view/inactive_view.dart
@@ -8,7 +8,7 @@ import 'package:gocast_mobile/views/video_view/video_player.dart';
 import 'package:gocast_mobile/base/networking/api/gocast/api_v2.pb.dart';
 
 class InactiveView extends ConsumerStatefulWidget {
-  final Int64? streamID;
+  final int? streamID;
 
   const InactiveView({
     super.key,
