+++ lib/views/course_view/list_courses_view/courses_list_view.dart
@@ -72,7 +72,7 @@ class CoursesList extends ConsumerWidget {
             courseId: course.id,
             semester:
                 course.semester.teachingTerm + course.semester.year.toString(),
-            lastLectureId: Int64(course.lastRecordingID),
+            lastLectureId: course.lastRecordingID,
             isCourse: true,
             onTap: () {
               Navigator.push(
